@@ -21,7 +21,7 @@
     "dependencies": {
         "@svgr/webpack": "^5.4.0",
         "clsx": "^1.0.4",
-        "jwt-decode": "^2.2.0",
+        "jwt-decode": "^3.0.0",
         "memoize-one": "^5.1.1",
         "oidc-client": "^1.10.1",
         "prop-types": "15.7.2",
@@ -40,6 +40,7 @@
         "react-intl": "^3.11.0"
     },
     "devDependencies": {
+        "is-equal": "1.6.1",
         "notistack": "^1.0.1",
         "@material-ui/core": "^4.9.5",
         "@material-ui/icons": "^4.5.1",
@@ -47,25 +48,25 @@
         "babel-eslint": "^10.1.0",
         "bufferutil": "^4.0.1",
         "eslint": "^7.9.0",
-        "eslint-config-prettier": "^2.9.0",
-        "eslint-config-react-app": "^5.2.1",
+        "eslint-config-prettier": "^8.0.0",
+        "eslint-config-react-app": "^6.0.0",
         "eslint-loader": "^4.0.2",
         "eslint-plugin-flowtype": "^5.2.0",
         "eslint-plugin-import": "^2.22.0",
         "eslint-plugin-jsx-a11y": "^6.3.1",
-        "eslint-plugin-prettier": "^2.6.0",
+        "eslint-plugin-prettier": "^4.0.0",
         "eslint-plugin-react": "^7.21.0",
         "eslint-plugin-react-hooks": "^4.1.2",
         "inferno": "^7.4.2",
         "nwb": "^0.25.x",
-        "prettier": "2.0.5",
-        "puppeteer": "^5.3.0",
-        "react": "^16.13.1",
-        "react-dom": "^16.13.1",
-        "react-intl": "^3.11.0",
+        "prettier": "^2.0.0",
+        "puppeteer": "^10.0.0",
+        "react": "^17.0.0",
+        "react-dom": "^17.0.0",
+        "react-intl": "^5.0.0",
         "react-router": "^5.1.2",
         "react-router-dom": "^5.1.2",
-        "type-fest": "^0.13.1",
+        "type-fest": "^2.0.0",
         "utf-8-validate": "^5.0.2",
         "webpack-plugin-serve": "^1.0.1"
     },
